AM65SW-126 drivers: pci: controller: dwc: pci-keystone: get PCIe reset GPIO earlier in probe
When this driver uses a GPIO as the PCIe reset signal and that GPIO
isn't ready for use in the first driver probe, it will hang when
linking the PCIe PHY on the next driver probe. To prevent this hang,
we move the getting of the GPIO to before the driver links the PCIe
PHYs to the device which allows the driver to cleanly exit with an
EPROBE_DEFER return. Since the timing of the use of the GPIO is
unchanged, this driver change has no known side effects.

Signed-off-by: Matt McKee <mmckee@phytec.com>